backend/requirements.txt
app==0.0.1
fastapi==0.116.1
pydantic==2.11.8
python-dotenv==1.1.1
Requests==2.32.5
uvicorn==0.35.0
python-multipart==0.0.20
firebase-admin==7.1.0
PyJWT==2.10.1
python-jose[cryptography]==3.5.0